% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pascal_row}
\alias{pascal_row}
\title{Adjusted Pascal's}
\usage{
pascal_row(r, odds_ratio = 1)
}
\arguments{
\item{r}{The row from pascals triangle to output}

\item{odds_ratio}{The odds ratio to use. Default is 1:1 but for events with probability of 25 percent, for example, use odds ratio 1:3}
}
\value{
A row from Pascals triangle
}
\description{
Output a row from an adjusted pascals triangle
}
\examples{
pascal_row(10);
pascal_row(10,3);
}
